make using DynamicalSystemsBase work
for now I have commented out all code that uses OrdinaryDiffEq
