+++ src/DynamicalSystemsBase.jl
@@ -5,7 +5,6 @@ Definitions of core system and data types used
 in in the ecosystem of DynamicalSystems.jl
 """
 module DynamicalSystemsBase
-using StaticArrays
 
 include("dataset.jl")
 include("reconstruction.jl")
@@ -14,7 +13,7 @@ include("neighborhoods.jl")
 
 include("dynamicalsystem/dynamicalsystem.jl")
 include("dynamicalsystem/discrete.jl")
-include("dynamicalsystem/continuous.jl")
+# include("dynamicalsystem/continuous.jl")
 include("dynamicalsystem/famous_systems.jl")
 
 # include("deprecations.jl")
