--- test/REQUIRE
@@ -1,1 +1,0 @@
-DiffEqCallbacks
